LayoutTests/fast/spatial-navigation/snav-search-optimization-expected.txt

1	2	3
4	
	6
7	8	9
PASS internals.lastSpatialNavigationCandidateCount() is 9
PASS internals.lastSpatialNavigationCandidateCount() is 6
PASS internals.lastSpatialNavigationCandidateCount() is 2
PASS internals.lastSpatialNavigationCandidateCount() is 9
PASS internals.lastSpatialNavigationCandidateCount() is 6
PASS internals.lastSpatialNavigationCandidateCount() is 2
PASS internals.lastSpatialNavigationCandidateCount() is 5
PASS internals.lastSpatialNavigationCandidateCount() is 2
